moved functions from `utilities` to `tools` and changed all (hopefully) associated directory paths
