+++ tools/Barrier3D_Plotting_Functions.py
